src/requirements.txt
Flask==3.0.3
kafka_python==2.0.2
pymongo==4.7.1
